+++ npm/lib/postinstall.js
@@ -16,7 +16,7 @@ if (forceInstall) {
   console.log("--force, ignoring caches");
 }
 
-const VERSION = "v0.11.0";
+const VERSION = "v0.12.0";
 const BIN_PATH = path.join(__dirname, "../bin");
 
 process.on("unhandledRejection", (reason, promise) => {